fix: expose log session helper in main
